fix(ops): knockout_pool int→bigint 승격 — KO 풀 산술 오버플로 근본 차단 (#226)
ops_live_stats.knockout_pool 이 1c 에서 int 로 정의돼 recompute 의
(v_total_buyins * v_bounty_cost)::int 가 bounty 1억 × total_buyins(≥22)
= 22억 > int max(21.4억) 에서 22003 numeric overflow 로 터지고, DEFERRED
트리거 커밋 시점에 전 참가자 변이가 막히는 문제. prize_pool 은 이미 bigint
인데 knockout_pool 만 int 인 비대칭이 근원.

- 마이그 20260706100000: 컬럼 int→bigint + recompute 재정의(live 정의 베이스,
  DECLARE bigint·산식 ::int 다운캐스트 제거 두 곳만, 본체 전건 보존)
- pgTAP ops_knockout_pool_bigint: total_buyins=22·bounty=1억 → 22억 실값을
  직접 recompute·DEFERRED 트리거 두 경로로 검증(RED: int→22003, GREEN: bigint→PASS)
- ops_live_stats_deferred.test: knockout_pool 단언 40000→40000::bigint
  (bigint 컬럼 대응, prize_pool 단언 컨벤션 정합)

bounty_cost CHECK 상한 1억은 무해 sanity bound 로 유지(별도 논의).

검증: 전체 pgTAP 49파일 521 PASS · quality EXIT0 · jest 4788 PASS

+++ uniqn-mobile/supabase/tests/ops_live_stats_deferred.test.sql
@@ -74,7 +74,7 @@ UPDATE public.ops_tournaments SET bounty_cost = 10000
   WHERE id = (current_setting('ops.t_id'))::uuid;
 SELECT is(
   (SELECT knockout_pool FROM public.ops_live_stats WHERE tournament_id = (current_setting('ops.t_id'))::uuid),
-  40000, 'tournaments 트리거 발화: knockout_pool=(3+1)×10000=40000');
+  40000::bigint, 'tournaments 트리거 발화: knockout_pool=(3+1)×10000=40000');
 
 -- ── [9] tournaments 비용 변경 트리거: buy_in_cost 변경 → prize_pool 재계산 ──
 UPDATE public.ops_tournaments SET buy_in_cost = 60000
